+++ lib/active_model/serializer/adapter/json_api/api_objects.rb
@@ -1,0 +1,13 @@
+module ActiveModel
+  class Serializer
+    module Adapter
+      class JsonApi
+        module ApiObjects
+          extend ActiveSupport::Autoload
+          autoload :Relationship
+          autoload :ResourceIdentifier
+        end
+      end
+    end
+  end
+end